+++ app/models/Centre.java
@@ -6,7 +6,7 @@ public class Centre extends BaseModel {
 
     private String name;
     private String code;
-    private Integer university_id;
+    private Long university_id;
     private Date deleted_at;
     private Date updated_at;
 
@@ -26,11 +26,11 @@ public class Centre extends BaseModel {
         this.code = code;
     }
 
-    public Integer getUniversityId() {
+    public Long getUniversityId() {
         return university_id;
     }
 
-    public void setUniversityId(Integer universityId) {
+    public void setUniversityId(Long universityId) {
         this.university_id = universityId;
     }
