+++ app/models/City.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class City extends BaseModel {
 
     private String name;
-    private Integer country_id;
+    private Long country_id;
     private Date deleted_at;
     private Date updated_at;
 
@@ -17,11 +17,11 @@ public class City extends BaseModel {
         this.name = name;
     }
 
-    public Integer getCountryId() {
+    public Long getCountryId() {
         return country_id;
     }
 
-    public void setCountryId(Integer country_id) {
+    public void setCountryId(Long country_id) {
         this.country_id = country_id;
     }
